[TSA2] added the sewer level
The sewer level is accessible from the hole in the debug area.
The sewer level's hole is the new entrance to the black market.
There are now less shops in the black market.
There are now sewer chests and ultra rare great chests in the sewer. Sewer chests can contain supplies and guns, and great chests can contain any random item.
The sewer has a new terrain generation type.
The player is slightly faster by default.